unified used of build_spark_submit
